+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## Unreleased
+- Breaking changes to commandline arguments. Switched over to Clap crate for parsing arguments which will help handle additional features.
 - Set priority of no date to 80
 - Add `--complete` feature
+- Update dependencies
 
 ## 2021-09-07 0.1.2
 - Color the item content based on the priority assigned